# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: Keepgoing API
  version: 1.0.0
  description: |
    블로그 + RAG + 이력서 생성 서비스의 백엔드 API 스펙.
    
    ## 주요 기능
    - 인증/인가 (Auth): JWT 기반 인증
    - 사용자 관리 (User): 프로필 관리, 관리자 기능
    
    ## 인증
    브라우저 환경에서는 HttpOnly 쿠키 기반 JWT 인증을 사용합니다.
    로그인/OAuth 성공 후 서버는 access/refresh 토큰을 쿠키로 설정하며, 클라이언트는 이후 요청에서 쿠키 기반 인증을 사용합니다.
    비브라우저 클라이언트와 수동 테스트에서는 `Authorization: Bearer <token>` 헤더 기반 접근도 사용할 수 있습니다.
    
    ## CSRF 보안
     이 서비스는 브라우저 환경에서 쿠키 기반 인증을 사용할 때 다음과 같은 전략으로 CSRF 위험을 완화합니다.
     - 인증 쿠키는 `HttpOnly`, `Secure`, `SameSite=Lax`(또는 더 엄격한) 옵션으로 설정됩니다.
     - 상태를 변경하는 모든 API는 `GET` 대신 `POST`, `PUT`, `PATCH`, `DELETE` 메서드를 사용합니다.
     - 서버는 별도의 CSRF 토큰 검증 매커니즘을 사용하지 않으므로, 위 정책과 API 설계를 전제로 보안 위협 모델을 구성합니다.
     - 브라우저 클라이언트는 다른 출처에서 이 API를 임베드하여 자동으로 쿠키가 전송되는 패턴을 피해야 합니다.

  contact:
    name: API Support
    email: api@blingbling.com

  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: 로컬 개발 서버
  - url: https://api.blingbling.com/v1
    description: 프로덕션 서버
...
paths:
...
  /auth/signup:
    post:
      tags: [Auth]
      summary: 회원가입
      operationId: signUp
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
            example:
              email: user@example.com
              name: 홍길동
              password: SecurePass123!
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignupResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: user@example.com
              password: SecurePass123!
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/me:
    get:
      tags: [Auth]
      summary: 현재 로그인 사용자 조회
      operationId: getMyInfo
      security:
        - BearerAuth: []
      responses:
        '200':
          description: 현재 로그인 사용자 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyInfoResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 액세스 토큰
...
  responses:
    BadRequest:
      description: 잘못된 요청 (유효성 검사 실패)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: VALIDATION_FAILED
              message: 입력값 검증 실패
              details:
                - field: email
                  message: 올바른 이메일 형식이 아닙니다
            timestamp: "2024-01-01T00:00:00Z"
            path: /auth/signup
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: AUTH_TOKEN_INVALID
              message: 유효하지 않은 토큰입니다
            timestamp: "2024-01-01T00:00:00Z"
            path: /users/me
...
    NotFound:
      description: 리소스를 찾을 수 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: USER_NOT_FOUND
              message: 사용자를 찾을 수 없습니다
            timestamp: "2024-01-01T00:00:00Z"
            path: /users/123
    Conflict:
      description: 리소스 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: USER_ALREADY_EXISTS
              message: 이미 존재하는 이메일입니다
            timestamp: "2024-01-01T00:00:00Z"
            path: /auth/signup
    TooManyRequests:
      description: 요청 횟수 초과
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: 제한 횟수
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: 남은 요청 횟수
        X-RateLimit-Reset:
          schema:
            type: integer
          description: 리셋 시간 (Unix timestamp)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: RATE_LIMIT_EXCEEDED
              message: 요청 횟수가 초과되었습니다. 잠시 후 다시 시도해주세요
            timestamp: "2024-01-01T00:00:00Z"
            path: /auth/login
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: 요청 성공 여부
        data:
          type: object
          nullable: true
          description: 성공 시 응답 데이터
        error:
          $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: string
          format: date-time
          description: 응답 시간
        path:
          type: string
          description: 요청 경로
        requestId:
          type: string
          description: 요청 추적 ID
          example: "req_1234567890"
      required: [success, timestamp, path]
    ErrorDetail:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: 에러 메시지
        field:
          type: string
          nullable: true
          description: 유효성 검사 실패 필드
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
          nullable: true
          description: 상세 에러 정보
      required: [code, message]
    ErrorCode:
      type: string
      enum:
        # Auth errors
        - AUTH_INVALID_CREDENTIALS
        - AUTH_TOKEN_EXPIRED
        - AUTH_TOKEN_INVALID
        - AUTH_REFRESH_TOKEN_INVALID
        - AUTH_EMAIL_NOT_VERIFIED
        # User errors
        - USER_NOT_FOUND
        - USER_ALREADY_EXISTS
        - USER_SUSPENDED
        - USER_DELETED
        # Validation errors
        - VALIDATION_FAILED
        - INVALID_INPUT
        - INVALID_PASSWORD_FORMAT
        - PASSWORD_TOO_WEAK
        # Permission errors
        - PERMISSION_DENIED
        - ROLE_INSUFFICIENT
        # Rate limit
        - RATE_LIMIT_EXCEEDED
        # System errors
        - INTERNAL_SERVER_ERROR
        - SERVICE_UNAVAILABLE
        - RESOURCE_NOT_FOUND
...
    SignUpRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          description: 사용자 이메일
        name:
          type: string
          minLength: 2
          maxLength: 50
          description: 사용자 이름
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          pattern: '^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$'
          description: |
            비밀번호 (최소 8자, 대소문자, 숫자, 특수문자 포함)
      required: [email, name, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required: [email, password]
    SignupResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                email:
                  type: string
                  format: email
                name:
                  type: string
              required: [id, email, name]
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                userId:
                  type: integer
                  format: int64
                email:
                  type: string
                  format: email
              required: [userId, email]
    MyInfoResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                userId:
                  type: integer
                  format: int64
                email:
                  type: string
                  format: email
                name:
                  type: string
                role:
                  $ref: '#/components/schemas/UserRole'
              required: [userId, email, name, role]

    # ========== User Schemas ==========
    UserRole:
      type: string
      enum: [USER, ADMIN]
      description: 사용자 권한
...